web: move package 'models' to 'api/'
