+++ internal/web/api/days.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/ShoshinNikita/budget-manager/internal/db"
 	"github.com/ShoshinNikita/budget-manager/internal/pkg/request_id"
-	"github.com/ShoshinNikita/budget-manager/internal/web/models"
+	"github.com/ShoshinNikita/budget-manager/internal/web/api/models"
 	"github.com/ShoshinNikita/budget-manager/internal/web/utils"
 )
