+++ internal/web/handlers_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/ShoshinNikita/budget-manager/internal/db"
 	"github.com/ShoshinNikita/budget-manager/internal/db/pg"
 	"github.com/ShoshinNikita/budget-manager/internal/pkg/request_id"
-	"github.com/ShoshinNikita/budget-manager/internal/web/models"
+	"github.com/ShoshinNikita/budget-manager/internal/web/api/models"
 )
 
 const (
